Avoid MySQL deadlock on EVENT_WAIT_INSTANCE

+++ mdw-services/src/com/centurylink/mdw/service/data/process/EngineDataAccessDB.java
@@ -671,7 +671,14 @@ public class EngineDataAccessDB extends CommonDataAccess implements EngineDataAc
     }
 
     public void removeEventWaitForProcessInstance(Long processInstanceId) throws SQLException {
-        String query = "delete from EVENT_WAIT_INSTANCE " +
+        String query;
+        if (db.isMySQL()) // This is to separate the row locking from delete statement of impacted rows, to avoid deadlock in MySQL
+            query = "delete E1 from EVENT_WAIT_INSTANCE E1 " +
+                "join EVENT_WAIT_INSTANCE E2 using (EVENT_WAIT_INSTANCE_ID) " +
+                "where E2.EVENT_WAIT_INSTANCE_OWNER_ID in " +
+                "(select ACTIVITY_INSTANCE_ID from ACTIVITY_INSTANCE where PROCESS_INSTANCE_ID=?)";
+        else
+            query = "delete from EVENT_WAIT_INSTANCE " +
                 "where EVENT_WAIT_INSTANCE_OWNER_ID in " +
                 "  (select ACTIVITY_INSTANCE_ID from ACTIVITY_INSTANCE where PROCESS_INSTANCE_ID=?)";
         db.runUpdate(query, processInstanceId);
